Makefile libft.h and fixes

+++ ft_bzero.c
@@ -1,6 +1,8 @@
+#include "libft.h"
+
 void	ft_bzero(void *s, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (n > 0)
